moved files into components folder
