--- src/Profile.tsx
@@ -1,13 +1,0 @@
-function Profile() {
-    return (
-
-        <div>
-         <h1>hi</h1>
-
-
-        </div>
-
-    );
-}
-
-export default Profile
